requirements.txt
selene~=2.0.0rc9
pytest~=8.3.4
Appium-Python-Client~=4.4.0
python-dotenv==0.20.0
allure-pytest==2.13.5
requests~=2.32.3
selenium~=4.27.1
allure-python-commons~=2.13.5
utils~=1.0.2
python-dotenv~=1.0.1
pydantic-settings==2.6.1











